teams-bot/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "teams-bot"
version = "0.1.0"
description = "Microsoft Teams meeting bot — joins meetings, starts transcriptions, retrieves events via Graph API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "msgraph-sdk>=1.5",
    "msgraph-beta-sdk>=1.5",
    "azure-identity>=1.17",
    "pydantic-settings>=2.6",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx",
]

[tool.hatch.build.targets.wheel]
packages = ["src/teams_bot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
